GO-177: Moved UserResponse to commons
