+++ Backend/go-team-service/src/test/java/com/game/on/go_team_service/TeamServiceTest.java
@@ -5,7 +5,6 @@ import com.game.on.go_team_service.config.CurrentUserProvider;
 import com.game.on.go_team_service.exception.BadRequestException;
 import com.game.on.go_team_service.exception.ConflictException;
 import com.game.on.go_team_service.exception.ForbiddenException;
-import com.game.on.go_team_service.exception.NotFoundException;
 import com.game.on.go_team_service.team.dto.*;
 import com.game.on.go_team_service.team.mapper.TeamMapper;
 import com.game.on.go_team_service.team.metrics.TeamMetricsPublisher;
@@ -14,7 +13,6 @@ import com.game.on.go_team_service.team.repository.TeamInviteRepository;
 import com.game.on.go_team_service.team.repository.TeamMemberRepository;
 import com.game.on.go_team_service.team.repository.TeamRepository;
 import com.game.on.go_team_service.team.service.TeamService;
-import com.game.on.go_user_service.dto.UserResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
